fix(top-rated): avoid PostgreSQL numeric overflow in weighted sort
Compute Bayesian ranking in double precision for ORDER BY so vote_count * vote_average intermediates do not overflow numeric(5,2).

+++ src/MovieApp.Infrastructure/Persistence/Search/SearchQueryBuilder.cs
@@ -282,14 +282,16 @@ internal static class SearchQueryBuilder
         decimal catalogMeanVoteAverage,
         int minimumVoteConfidence)
     {
-        var minimumVotes = (decimal)minimumVoteConfidence;
-        var catalogMean = catalogMeanVoteAverage;
+        // Use double precision for SQL ordering. PostgreSQL numeric(5,2) intermediates overflow
+        // when vote_count is multiplied by vote_average during Bayesian score translation.
+        var minimumVotes = (double)minimumVoteConfidence;
+        var catalogMean = (double)catalogMeanVoteAverage;
 
         return ApplyDeterministicTieBreak(
             query
                 .OrderByDescending(item =>
-                    ((decimal)item.VoteCount / (item.VoteCount + minimumVotes)) * item.VoteAverage
-                    + (minimumVotes / (item.VoteCount + minimumVotes)) * catalogMean)
+                    item.VoteCount / (item.VoteCount + minimumVotes) * (double)item.VoteAverage
+                    + minimumVotes / (item.VoteCount + minimumVotes) * catalogMean)
                 .ThenByDescending(item => item.VoteCount)
                 .ThenBy(item => item.Title));
     }
